Expose the evaluation result
We recently updated the aiken crate to expose the underlying Evaluation Result; this has the benefit of exposing traces even when the contract succeeds! This updates the code to start that process. I'd also like to update the errors to throw a rich error type in the case of failure, but I'll crack one egg at a time